Add full test files for 2020

+++ src/main/java/com/cairns/rich/aoc/_2020/Day13.java
@@ -26,7 +26,7 @@ class Day13 extends Base2020 {
         minBus = bus;
       }
     }
-    return minBus * (min - earliest);
+    return (int) (minBus * (min - earliest));
   }
 
   @Override
